+++ app/routers/om/profile.py
@@ -24,7 +24,7 @@ async def create_profile(
     commons: CommonsDep,
     initial_node_id: int = Body(..., description="Initial node ID", examples=[35]),
     final_node_id: int = Body(..., description="Final node ID", examples=[38]),
-    middle_nodes: Optional[List[int]] = Body(None, description="Middle nodes IDs", examples=[[36, 37]]),
+    middle_nodes: Optional[List[int]] = Body(None, description="Middle nodes IDs", examples=[[37]]),
     links_distance: int = Body(..., description="Links distance", examples=[1]),
     scale_eh: int = Body(..., description="Scale EH", examples=[1000]),
     scale_ev: int = Body(..., description="Scale EV", examples=[1000]),